Fix/Change: (hyperdrive--format-entry-url) Accept an entry struct
The previous version of this function hyperdrive--format-url depended
on the persisted hyperdrive--namespaces variable, which we no longer
use.

hyperdrive--format-entry-url now accepts an entry struct instead of a
url, since we need to access the alias of the hyperdrive and all the
functions which call it already have access to an entry. If this
function instead accepted a url, we would have to use gethash to find
the hyperdrive.