Fix: (hyperdrive-fill) Set :noquery t
Allow the process buffer to be killed without prompting the user.  (It
seems like this shouldn't be necessary, but there appears to be a race
condition or "Schrödinger's process" problem in which sometimes, even
when `process-live-p' returns nil for a process, calling `kill-buffer'
on its buffer causes Emacs to prompt the user anyway, so we set this
option to prevent that.  It's more correct to do so, anyway, since
this request can be aborted harmlessly.)