Fix (hyperdrive-complete-hyperdrive): Don't truncate public key
Using an alist of cells like
(<first 6 chars of public-key> . <entry>)
breaks when the first 6 chars of two public keys match.

We'd figured the odds were miniscule, but it's occurring
now when user calls `hyperdrive-open', pastes in a URL,
and then changes the last few characters of it.

In this commit, display full public keys and aliases without the
hyper:// prefix and display duplicate candidates for writable drives:
both the alias and public key.