Change: Remove with-version-p from hyperdrive-entry etc slot
This commit removes with-version-p. Now, everywhere we want to
determine whether or not to use a version number, we look at the
version slot of hyperdrive-entry. The version slot corresponds to the
presence or absence of a version number in the entry URL.

Additionally, this commit adds a version-last-modified slot to
hyperdrive-entry, which records the ETag from hyper-gateway.
version-last-modified corresponds to the hyperdrive version at which
the entry was last modified before the current version. For directory
entries, this will always equal the version slot, unless the version
slot is nil, in which case version-last-modified will equal the latest
version of the hyperdrive.