Remove: (hyperdrive-save-buffer) Remove unnecessary function
The original design was modeled after save-buffer and write-file. Now
that write-contents-functions maps to hyperdrive--write-contents,
save-buffer behaves as expected (saves the current hyperdrive entry
buffer to its URL). To save a non-hyperdrive buffer as a hyperdrive
entry or to save the current hyperdrive entry buffer to a different
URL, use hyperdrive-write-buffer.