Change: (hyperdrive-history) Use range-entries
Each ewoc entry is a now range-entry's car is a range plist,
containing existsp (one of nil, t, or unknown), and its cdr is a
hyperdrive entry.

By storing the range-end value in each ewoc entry, we can easily
determine the end of the range of unknown entries, which are not
stored in hyperdrive-version-ranges. The alternative is to iterate
through hyperdrive-version-ranges to find the range-start of the range
which follows, and then decrement it to get the range-end. I think
this way is simpler.