Change: (hyperdrive-diff-file-entries) Remove spurious ELSE
The ELSE handler fired when either the request for OLD-ENTRY or
NEW-ENTRY failed. However, in normal usage, the request for one of the
entries will fail when at the beginning or end of an existent range.

Only when both requests fail, should we call ELSE, but plz.el queues
do not yet support that.