Add: (hyperdrive-mirror.el) Add revert-buffer-function
Instead of checking that a mirror query has already been completed,
store the prior query and revert the mirror buffer when the upload
finishes, showing that all files now have the "same" status.

Reverting the buffer calls hyperdrive-mirror again, checking for
changes in the local filesystem as well as the hyperdrive.