WIP: (-fill-version-ranges) Fill existent and nonexistent ranges
Two codepaths:

- fill existent entries in series
- fill nonexistent entries in parallel

Filling existent entries is logically recursive: fill entry, get its
range-start, then fill the entry at the range-end of the prior range.

Filling nonexistent entries is not recursive: In parallel, send
a request to every version of entry between

(hyperdrive-entry-version entry)
and
(- (hyperdrive-entry-version entry) hyperdrive-queue-size)

If the earliest requested version is existent, go back to the
"existent series" codepath. Otherwise, repeat the "nonexistent
parallel" codepath.

So good, so far.

The problem is that the finally callback passed to
`hyperdrive-fill-version-ranges' is not called in the correct place.
We check the earliest requested version at the end of the "nonexistent
parallel" codepath inside a separate queue finalizer for the codepath.
Since the "nonexistent parallel" async requests are not made in the
main queue used by "existent series" codepath, the main queue runs
whenever the "existent series" codepath hands control over to the
"nonexistent parallel" codepath.

What about multiple queues per request, each with its own finalizer?