Revert: -view-file-other-window
Because view-buffer calls switch-to-buffer, it would be awkward to
override how it displays the buffer, so it may not be worthwhile to
provide -view-file-other-window commands.  So removing this for now;
it can be squashed away, or we can decide to hack up a workaround and
add a -view-file-other-window command for the history buffer too.