Meta: Track pre-commit hook to build the manual
It is necessary to check git status since we only want to stage
hyperdrive.texi when hyperdrive.org is also staged.  If instead we
relied on GNU Make's built-in timestamp comparison, `git add
doc/hyperdrive.texi` would run when doc/hyperdrive.org is modified in
the worktree but not among the files staged for commit.