Change: (hyperdrive-menu--entry) New convenience function
This doesn't have to be a macro as suggested in the todo item, which
this addresses.

However, the order matters in which we try `transient--prefix'
and `transient-current-prefix', and use the first that is non-nil.
`transient-current-prefix' may be another prefix from which we invoke
the current prefix, so `transient--prefix' has to take precedence.