Change: (-menu-set-nickname) No longer explicitly refresh transient
Since the previous commit that isn't necessary anymore as the
transient buffer is now refreshed after every suffix command.  Also,
transient prefix commands are only intended for interactive use, so
this resulted in an error, and if this command is invoked directly,
then we likely wouldn't want that to bring up the transient.