Change: Replace hyperdrive--org- prefix with hyperdrive-org--
A few symbols already used the latter prefix and there was a TODO
item about having to decide which prefix should be used.  Since we
are also going to use shorthands, it IMO makes more sense to use the
latter prefix.  The "ho-" shorthand prefix makes more sense if the
two affected prefixes are "hyperdrive-org-" and "hyperdrive-org--".