Fix: Fix indentation
Unfortunately the indentation code is not smart enough to realize,
that a list that begins with a keyword cannot possibly be a function
call and should therefore not be indented as such.  (Vectors and
arglists are also indented as if they were function calls.)

Emacs-devel opposes the idea of fixing these "edge-cases", and instead
suggests that this should be worked around by putting a space between
the opening parenthesis and the first element.