Chore: Set one variable per setf and setq-local form
Especially with setf, this helps the reader immediately see what is
a place and what a value.  Given something like:

  (setf ...
        (alist-get 'foo alist)
        (alist-get 'bar alist)
        ...)

It is not possible to know if "foo" is being set to "bar", or whether
they are unrelated, without looking at the surrounding and possibly
even having to count arguments.  To make it possible to rely on this
being done in hairy cases -- without having to double-check -- it is
important that this is done consistently, i.e., even if the simple
cases where it is "not necessary".

This also makes it easier to remove or comment out some assignment,
including the ones at the beginning and the end, though that might
be more useful in user configuration than a stable code-base.