Chore: Begin multi-line values of defvar on a new line
Unlike many other commits in this series, this isn't about long
lines per se (though it can).  I find the left whitespace of random
width (depending on the variable name), a bit awkward.  While to my
knowledge, beginning the value on a new line, is the norm, if it does
not fit on the same line as the variable.