Chore: Join some short lines
Given the general direction of this commit series, this might come as
a bit of a surprise, but I am in fact not a short-line absolutist. ;)

Many of these are instances were we (or so I assume) added line-breaks
to shorten long lines in the past.  Since we have switched to using
symbol shorthands, this isn't necessary in these cases anymore.