Chore: Use semantically correct contitionals
Use `and', `or' and `and-let*' when the return value matters.

Use `when', `unless', `when-let' and `when-let*' when the return
value does not matter -- only the side-effect.

`if' and `cond' may appear in either context, but `if' must always
have a ELSE part.

Sometimes the value matters and there is a side-effect.  In such
cases one should use `and' and `progn'.  That is uglier than just
using `when', but that is a good thing; it draws the readers
attention, who might otherwise have overlooked the fact that both
matter.

Co-authored-by: Joseph Turner <joseph@ushin.org>