Change: Remove hyperdrive-reuse-buffers
When implementing the uniquification code, it became clear that it
wasn't feasible to attempt to handle all of the edge cases where
multiple buffers could accidentally be created that visit the same
entry at different versions.

By removing the feature, we simplify the expected behavior and allow
users to kill extra buffers as they see fit.