Change: (h/start) Move default start logic to default start function
h/install-in-progress-p, h//hyper-gateway-ushin-path, and
h/gateway-process are specific to the default method of running
hyper-gateway-ushin as an Emacs subprocess, so they should be checked
in the default start function.

This commit also changes the behavior when the gateway is being
installed and there exists an old copy of the gateway in the PATH.
Previously an error was signaled, preventing the old gateway from
starting.  Now, the old gateway is started and a message appears,
"Gateway installation in-progress; starting old gateway anyway."