Change: (h/menu-bar-menu) Remove gateway status indicator
Since the menu bar's items can't be updated while the menu is
displayed, showing an indicator could be misleading to the user, since
it could become outdated while the menu is displayed (e.g.  while
starting the gateway).

So we won't show such indicators in the menu bar.  Instead, the
Transient ~hyperdrive-menu~ UI is a suitable "dashboard" for such
indicators, because the Transient can be refreshed while it's
displayed.

Co-authored-by: Adam Porter <adam@alphapapa.net>