Docs: Remove suggestion to reset persist variables
We originally added this suggestion when the two persisted variables
were still changing rapidly.  It seems unlikely now that bad persist
variables are the root of user issues, and we don't want users to
delete their data for no reason.  Also, if there is an issue with the
persist variables, we'd prefer for users keep the invalid data so they
can pass it along to the maintainers for diagnosis.