Change: Use new X-Drive-Version header
Replace references to ETag with X-Drive-Version as semantically
appropriate.  Now, when we want the latest drive version, we use the
new header, and when we want the sequence number of an entry, we use
ETag.

Since the new X-Drive-Version header is returned for HEAD/GET requests
to file entries (whereas the ETag header only ever returned the latest
drive version in certain cases), we can now consolidate the logic for
filling the latest version of a hyperdrive into the he/api function.