Fix: (h/menu-refresh) More reliably refresh transient
Previously, the transient refreshed only h/start was invoked from
within the same transient.  However, if h/start was invoked with M-x
or if the transient was closed and reopened, the gateway status would
hang at "starting".

Now, regardless of how h/start was invoked, the gateway status should
update correctly whenever the transient is open and not suspended,
such as when the minibuffer is active.

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>