Change: (he//invalidate) Only invalidate select entry slots
Previously, EWOC entry data was replaced entirely with the entry which
was available in the function which called he//invalidate, usually
he//api-then.  However, when h/forget-file was called, the entry it
passed to he/api and he//api-then lacked the ETC slot RANGE-END and
EXISTSP values, and so when the history buffer ewoc node entry was
invalidated, the ewoc printer signaled an error.

Now, ewoc node entries are reused, and only select slots are updated.
This ensures that RANGE-END and EXISTSP values remain.

If server responses always returned all metadata necessary to render
each entry, this change might not be necessary.  However range-end
data requires traversing backward through the version history and
doing this for each request may pose a performance issue.

Also, there are other metadata which are buffer-specific and should
not be erased when invalidating an ewoc node entry, such as the ETC
slot DISPLAY-NAME key.  By only refreshing specific slots, we are
likely to prevent confusing bugs in the future if other
buffer-specific data gets stored in buffer-local hyperdrive entries.