Change: (hyperdrive-menu) Change description to "..." from "starting"
When h//gateway-ready-p is nil but h/gateway-live-p is nil, the
gateway process is running but not responsive.  This state occurs when
the gateway is either starting or stopping, so this commit intends to
change the description for this state to something ambiguous which
makes sense immediately after starting or stopping the gateway.

Also this shorter description does not cause the "Bookmark" Transient
group to temporarily move to the right.