Change: (h/history-load-range) Don't use h/history-load-limit
Apparently, loading the history for certain versions automatically
loads the history for older version as well.  This means that small
gaps in the history may be filled unexpectedly.  This change is a
workaround to avoid this confusing UI in the base case where the user
attempts to load the entire history.

Previously, we'd end up with something like this, where extra bits of
version history were loaded unintentionally:

[public-key:1m51x54k…] /posts/agregore-mobile.html
 Exists  Drive Version Range    Size      Last Modified
    Yes              623-843    7.7k
    Yes              379-622      6k
     No                  378
Unknown                  377
     No              375-376
Unknown              373-374
     No              370-372
Unknown              359-369
     No                  358
Unknown                2-357
     No                    1

Now, the whole history is loaded, which takes longer, but at least
gives a consistent result:

[public-key:1m51x54k…] /posts/agregore-mobile.html
 Exists  Drive Version Range    Size      Last Modified
    Yes              623-843    7.7k
    Yes              379-622      6k
     No                1-378