Change: (he/exists-p) Refactor to request existence from gateway
The timeout may not be necessary, but since this function will
synchronously block h/menu and the menu bar, we want to ensure that it
returns quickly.

An alternative solution that I considered was to mark an entry as
"deleted" in hyperdrive-delete, "unsaved" in the createp block of
hyperdrive-find-file, and "existent" in hyperdrive-entry--api-then.
That solution would have required tracking another piece of state in
the hyperdrive-current-entry local variable, and would have presented
another barrier to integration with Magic File Names if we decide to
go down that route.

This solution implemented in this commit may potentially be slower,
but in my manual testing, h/menu and the menu bar open instantly on my
slow laptop.