Change: Refactor fons-relation struct; remove merge-relations
Previously, fons-relation structs were agnostic of the type of
relation they contained: source or blocker.  This meant that it was
necessary for the fons-relation struct to have an additional blocked-p
slot, and that the fons-relations function had to iterate over the
list of relations once more to fill that slot.  It also meant that, in
order to calculate edges for fons-view, hash tables containing
fons-relations for sources, blockers, and blocked had to be combined
into merge-relations, which was an alist mapping 'sources, 'blockers,
and 'blocked to each hash table.

Now, fons-relation structs contain separate paths for sources,
blockers, and blocked.  This change leads to simpler code (though with
exactly the same LOC!) and more consistent data model.